+++ script/bj_invoker.rb
@@ -1,4 +1,4 @@
-@routing = Workling::Starling::Routing::ClassAndMethodRouting.new
+@routing = Workling::Routing::ClassAndMethodRouting.new
 unnormalized = REXML::Text::unnormalize(STDIN.read)
 message, command, args = *unnormalized.match(/(^[^ ]*) (.*)/)
 options = Hash.from_xml(args)["hash"]