+++ script/listen.rb
@@ -1,8 +1,8 @@
 puts '=> Loading Rails...'
 
 require File.dirname(__FILE__) + '/../../../../config/environment'
-require File.dirname(__FILE__) + '/../lib/workling/starling/poller'
-require File.dirname(__FILE__) + '/../lib/workling/starling/routing/class_and_method_routing'
+require File.dirname(__FILE__) + '/../lib/workling/remote/invokers/poller'
+require File.dirname(__FILE__) + '/../lib/workling/routing/class_and_method_routing'
 
 puts '** Rails loaded.'
 puts '** Starting Workling::Starling::Poller...'
@@ -11,7 +11,7 @@ puts '** Use CTRL-C to stop.'
 ActiveRecord::Base.logger = Workling::Base.logger
 ActionController::Base.logger = Workling::Base.logger
 
-poller = Workling::Starling::Poller.new(Workling::Starling::Routing::ClassAndMethodRouting.new)
+poller = Workling::Starling::Poller.new(Workling::Routing::ClassAndMethodRouting.new)
 
 trap(:INT) { poller.stop; exit }