+++ script/starling_status.rb
@@ -3,15 +3,15 @@ require 'pp'
 puts '=> Loading Rails...'
 
 require File.dirname(__FILE__) + '/../config/environment'
-require File.dirname(__FILE__) + '/../vendor/plugins/workling/lib/workling/starling/poller'
-require File.dirname(__FILE__) + '/../vendor/plugins/workling/lib/workling/starling/routing/class_and_method_routing'
+require File.dirname(__FILE__) + '/../vendor/plugins/workling/lib/workling/remote/invokers/poller'
+require File.dirname(__FILE__) + '/../vendor/plugins/workling/lib/workling/routing/class_and_method_routing'
 
 puts '** Rails loaded.'
 
 trap(:INT) { exit }
 
 begin
-  client = Workling::Starling::Client.new
+  client = Workling::Clients::Starling.new
   client.reset
   
   client.stats # do this so that connection is shown as established below.