+++ test/starling_client_test.rb
@@ -2,8 +2,8 @@ require File.dirname(__FILE__) + '/test_helper'
 
 context "The starling client" do
   specify "should be able to connect to multiple Starling instances" do
-    Workling::Starling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345, 127.0.0.1:12346", :memcache_options => { :namespace => "myapp_development" } }
-    client = Workling::Starling::Client.new
+    Workling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345, 127.0.0.1:12346", :memcache_options => { :namespace => "myapp_development" } }
+    client = Workling::Clients::Starling.new
     
     client.starling_urls.should.equal ["localhost:12345", "127.0.0.1:12346"]
     client.connection.servers.first.host.should == "localhost"
@@ -14,8 +14,8 @@ context "The starling client" do
   end
   
   specify "should load it's config as well as any given MemCache options from RAILS_ENV/config/starling.yml" do
-    Workling::Starling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345", :memcache_options => { :namespace => "myapp_development" } }
-    client = Workling::Starling::Client.new
+    Workling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345", :memcache_options => { :namespace => "myapp_development" } }
+    client = Workling::Clients::Starling.new
     
     client.starling_urls.should.equal ["localhost:12345"]
     client.connection.servers.first.host.should == "localhost"
@@ -25,8 +25,8 @@ context "The starling client" do
   
   
   specify "should load it's config correctly if no memcache options are given" do
-    Workling::Starling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345" }
-    client = Workling::Starling::Client.new
+    Workling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345" }
+    client = Workling::Clients::Starling.new
 
     client.starling_urls.should.equal ["localhost:12345"]
   end